+++ CMakeLists.txt
@@ -7,4 +7,4 @@ add_executable(42Cursus test/main.c libft/libft.h libft/ft_atoi.c libft/ft_bzero
 add_executable(printff printf/ft_printf.h printf/ft_printf.h printf/ft_conversions.c printf/ft_conversions_h.c printf/ft_putc.c printf/ft_putx.c printf/ft_putn.c printf/ft_putp.c printf/ft_puts.c printf/ft_putu.c printf/ft_printf.c printf/ft_itoa.c printf/ft_utoa.c printf/ft_xtoa.c printf/ft_libft.c)
 add_executable(getNextLine getNextLine/get_next_line.h getNextLine/get_next_line.c getNextLine/file getNextLine/get_next_line_utils.c)
 add_executable(getNextLine_bonus getNextLine/file getNextLine/file_bonus getNextLine/get_next_line_bonus.c getNextLine/get_next_line_utils_bonus.c getNextLine/get_next_line_bonus.h)
-add_executable(push_swap pushSwap/push_swap.h pushSwap/push_swap.c pushSwap/ft_s.c pushSwap/ft_p.c pushSwap/ft_rr.c pushSwap/ft_r.c pushSwap/ft_stack.c pushSwap/ft_sort.c pushSwap/ft_find.c pushSwap/ft_lst.c)
+add_executable(push_swap pushSwap/push_swap.h pushSwap/push_swap.c pushSwap/ft_s.c pushSwap/ft_p.c pushSwap/ft_rr.c pushSwap/ft_r.c pushSwap/ft_stack.c pushSwap/ft_sort.c pushSwap/ft_find.c pushSwap/ft_lst.c pushSwap/ft_libft.c pushSwap/ft_place.c)
