+++ pushSwap/ft_s.c
@@ -6,7 +6,7 @@
 /*   By: emaydogd <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/27 12:50:10 by emaydogd          #+#    #+#             */
-/*   Updated: 2023/06/05 00:10:47 by emaydogd         ###   ########.fr       */
+/*   Updated: 2023/06/09 18:48:03 by emaydogd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include "push_swap.h"
@@ -31,7 +31,6 @@ void	sa(t_stack *a)
 
 void	sb(t_stack *b)
 {
-
 	s(b);
 	printf("sb\n");
 }
