+++ pushSwap/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: emaydogd <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/27 12:44:07 by emaydogd          #+#    #+#             */
-/*   Updated: 2023/06/05 02:23:23 by emaydogd         ###   ########.fr       */
+/*   Updated: 2023/06/10 16:09:02 by emaydogd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #ifndef PUSH_SWAP_H
@@ -41,6 +41,14 @@ typedef struct s_stack {
 	t_node	*stack;
 }	t_stack;
 
+typedef struct s_optimize {
+	int	ra;
+	int	rb;
+	int	rr;
+	int	rra;
+	int	rrb;
+	int	rrr;
+}	t_optimize;
 
 void	sa(t_stack *a);
 void	sb(t_stack *b);
@@ -70,15 +78,21 @@ t_node	*ft_find_min(t_stack a);
 t_node	*ft_find_max(t_stack a);
 t_node	*ft_find_next(int val, t_stack s);
 t_node	*ft_find_prev(int val, t_stack s);
+void	ft_check_max(t_stack b, int val, int *lst, int j);
+int		*ft_place_new_middle(t_stack b, int val, int *lst, int j);
+void	ft_exec(int *lst, t_stack *a, t_stack *b);
 
 void	ft_place_new_min(t_stack b, int *lst, int j);
 void	ft_place_new_max(t_stack b, int *lst, int j);
 void	ft_rotate_end(t_stack b, int *lst, int j);
+void	ft_initopt(t_optimize	*opt);
 
 int		ft_lstlen(const int *lst);
 void	ft_lstprint(const int *lst);
 void	ft_lstreset(int *lst);
 int		*ft_lstoptimize(int *lst);
 int		ft_getval(t_stack a, int i);
+
+int		ft_atoi(const char *str);
 
 #endif //PUSH_SWAP_H
